+++ CHANGES.md
@@ -5,6 +5,7 @@ Version 0.7.0 (Pending)
   previously created in a manual step)
 - Fixed redirect bug when running behind load balancer that's using HTTPS
 - Fixed NullPointerException when a request has no cookies
+- Added Attendee Detail screen, add note functionality
 
 Version 0.6.0 (5/25/2019)
 - Added Till Detail Report after discussion w/ Treasury Dept. This report will be included with each